Upgrade bootstrap version to 2.3.2. It is no longer a supported version however it at least fixes some bugs with the version we had.
This also moves all of the pieces of code that we changed in the main bootstrap framework into the customcss.css file. By doing this we should be able to move to bootstrap 3 slightly easier if we ever decide to. Since we are now actually using the stock boostrap framework and applying changes to the top of it, we can now easily use the minified css, which was changed in main.css.

This fixes #8.